move moviepage
